requirements.txt
Django==1.7.3
MySQL-python==1.2.5
django-admin-tools==0.5.2
django-filter==0.9.1
django-fsm==2.1.0
django-fsm-admin==1.1.0
django-selectable==0.9.0
djangorestframework==3.0.3
